requirements.txt
deepgram-sdk==3.4.0 # python > 3.10 needed for this to work
numpy==1.24.4
openai-whisper==20231117
python-dotenv==1.0.1
azure-cognitiveservices-speech==1.38.0
google-cloud-speech==2.27.0
assemblyai==0.31.0
faster-whisper==1.0.3
